Refactoring: GeoDataRiverSurvey::ShiftRiverPathCenterCommand is moved to separated files.
